Made most of the implementation for the ball class notes some not implementented exceptions. Waiting for further discussion to decide how to adjust the speed.
Also changed the point to be a float point to ensure no loss of data while using floating point velocities. Since OpenGL can handle either, it is to our advantage to use floats.

+++ Pong/DrawableObject.cpp
@@ -1,6 +1,6 @@
 #include "DrawableObject.h"
 
- DrawableObject::DrawableObject(int x, int y, float color[])
+ DrawableObject::DrawableObject(float x, float y, float color[])
 {
     pnt.x = x; 
     pnt.y = y;
